Refactor shared/modules/ssgcommon.py to create new /ssg module
 - Splits ssgcommon.py into multiple submodules
 - Points the build system at new ssg module location
 - Updates all references to use new ssg module
