+++ ssg/__init__.py
@@ -1,0 +1,9 @@
+import ssg._checks as checks
+import ssg._constants as constants
+import ssg._jinja as jinja
+import ssg._utils as utils
+import ssg._xml as xml
+import ssg._yaml as yaml
+
+# Compatibility shims
+import ssg._shims as shims
