+++ ssg/_constants.py
@@ -1,0 +1,59 @@
+import datetime
+import os.path
+
+JINJA_MACROS_DEFINITIONS = os.path.join(os.path.dirname(os.path.dirname(__file__)),
+    "shared", "macros.jinja")
+
+xml_version = """<?xml version="1.0" encoding="UTF-8"?>"""
+
+datastream_namespace = "http://scap.nist.gov/schema/scap/source/1.2"
+ocil_namespace = "http://scap.nist.gov/schema/ocil/2.0"
+oval_footer = "</oval_definitions>"
+oval_namespace = "http://oval.mitre.org/XMLSchema/oval-definitions-5"
+ocil_cs = "http://scap.nist.gov/schema/ocil/2"
+xccdf_header = xml_version + "<xccdf>"
+xccdf_footer = "</xccdf>"
+bash_system = "urn:xccdf:fix:script:sh"
+ansible_system = "urn:xccdf:fix:script:ansible"
+puppet_system = "urn:xccdf:fix:script:puppet"
+anaconda_system = "urn:redhat:anaconda:pre"
+cce_uri = "https://nvd.nist.gov/cce/index.cfm"
+stig_ns = "http://iase.disa.mil/stigs/Pages/stig-viewing-guidance.aspx"
+ssg_version_uri = \
+    "https://github.com/OpenSCAP/scap-security-guide/releases/latest"
+OSCAP_VENDOR = "org.ssgproject"
+OSCAP_DS_STRING = "xccdf_%s.content_benchmark_" % OSCAP_VENDOR
+OSCAP_GROUP = "xccdf_%s.content_group_" % OSCAP_VENDOR
+OSCAP_GROUP_PCIDSS = "xccdf_%s.content_group_pcidss-req" % OSCAP_VENDOR
+OSCAP_GROUP_VAL = "xccdf_%s.content_group_values" % OSCAP_VENDOR
+OSCAP_GROUP_NON_PCI = "xccdf_%s.content_group_non-pci-dss" % OSCAP_VENDOR
+XCCDF11_NS = "http://checklists.nist.gov/xccdf/1.1"
+XCCDF12_NS = "http://checklists.nist.gov/xccdf/1.2"
+min_ansible_version = "2.3"
+ansible_version_requirement_pre_task_name = \
+    "Verify Ansible meets SCAP-Security-Guide version requirements."
+
+oval_header = (
+    """
+<oval_definitions
+    xmlns="{0}"
+    xmlns:oval="http://oval.mitre.org/XMLSchema/oval-common-5"
+    xmlns:ind="{0}#independent"
+    xmlns:unix="{0}#unix"
+    xmlns:linux="{0}#linux"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://oval.mitre.org/XMLSchema/oval-common-5 oval-common-schema.xsd
+        {0} oval-definitions-schema.xsd
+        {0}#independent independent-definitions-schema.xsd
+        {0}#unix unix-definitions-schema.xsd
+        {0}#linux linux-definitions-schema.xsd">"""
+    .format(oval_namespace))
+
+timestamp = datetime.datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%S")
+
+PKG_MANAGER_TO_SYSTEM = {
+    "yum": "rpm",
+    "zypper": "rpm",
+    "dnf": "rpm",
+    "apt_get": "dpkg",
+}
