+++ ssg/_xml.py
@@ -1,0 +1,47 @@
+import platform
+
+from ssg._constants import *
+
+
+try:
+    from xml.etree import cElementTree as ElementTree
+except ImportError as e:
+    print(e)
+    import cElementTree as ElementTree
+
+def oval_generated_header(product_name, schema_version, ssg_version):
+    return xml_version + oval_header + \
+        """
+    <generator>
+        <oval:product_name>%s from SCAP Security Guide</oval:product_name>
+        <oval:product_version>ssg: %s, python: %s</oval:product_version>
+        <oval:schema_version>%s</oval:schema_version>
+        <oval:timestamp>%s</oval:timestamp>
+    </generator>""" % (product_name, ssg_version, platform.python_version(),
+                       schema_version, timestamp)
+
+def parse_xml_file(filename):
+    """
+    Given a filename, return the corresponding ElementTree
+    """
+    with open(filename, 'r') as xml_file:
+        filestring = xml_file.read()
+        tree = ElementTree.fromstring(filestring)
+    return tree
+
+
+def map_elements_to_their_ids(tree, xpath_expr):
+    """
+    Given an ElementTree and an XPath expression,
+    iterate through matching elements and create 1:1 id->element mapping.
+
+    Raises AssertionError if a matching element doesn't have the ``id`` attribute.
+
+    Returns mapping as a dictionary
+    """
+    aggregated = {}
+    for element in tree.findall(xpath_expr):
+        element_id = element.get("id")
+        assert element_id is not None
+        aggregated[element_id] = element
+    return aggregated
